Corrected error and added comments
